requirements.txt
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.28.0
plotly>=5.15.0
requests>=2.28.0
beautifulsoup4>=4.11.0
schedule>=1.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
scholarly>=1.7.0
lxml>=4.9.0
fake-useragent>=1.1.0
pybliometrics>=1.0.0  # Added for Scopus
urllib3>=1.26.0
